requirements.txt
supabase
psutil
requests
crawl4ai
python-dotenv
psutil